requirements.txt
Pillow
requests
PyYAML
